ComponentID is now size_t to be compatible with typeid().has_code

+++ Bulb/include/Bulb/ECS/ECSTypes.hpp
@@ -2,7 +2,7 @@
 
 namespace blb::ecs {
 	using EntityID		= uint32_t;
-	using ComponentID	= uint64_t;
+	using ComponentID	= size_t;
 
 	inline constexpr EntityID INVALID_ENTITY_ID = 0;
 }
